// Angular/src/app/models/suppliersDto.ts
export default class SuppliersDto {
     id: number = 0;
     name: string = "";
     phone: string = "";
     activity_time: string = "";
     minPrice: number = 0;
     maxPrice: number = 0;
     description: string = "";
     src: string = "";
     srcLogo: string = "";
     idCategory: number = 0;
     idArea: number = 0;
     idAddress: number = 0;
    
     
    

 
    
}